+++ smarts/core/vehicle.py
@@ -313,7 +313,22 @@ class Vehicle:
         return VEHICLE_CONFIGS[self._vehicle_config_type].vehicle_type
 
     @staticmethod
+    def agent_vehicle_dims(mission) -> BoundingBox:
+        if mission.vehicle_spec:
+            # mission.vehicle_spec.veh_config_type will always be "passenger" for now,
+            # but we use that value here in case we ever expand our history functionality.
+            vehicle_config_type = mission.vehicle_spec.veh_config_type
+            return BoundingBox.copy_with_defaults(
+                mission.vehicle_spec.dimensions,
+                VEHICLE_CONFIGS[vehicle_config_type].dimensions,
+            )
+        # non-history agents can currently only control passenger vehicles.
+        vehicle_config_type = "passenger"
+        return VEHICLE_CONFIGS[vehicle_config_type].dimensions
+
+    @classmethod
     def build_agent_vehicle(
+        cls,
         sim,
         vehicle_id,
         agent_interface,
@@ -327,29 +342,21 @@ class Vehicle:
     ):
         mission = mission_planner.mission
 
-        if mission.vehicle_spec:
-            # mission.vehicle_spec.veh_config_type will always be "passenger" for now,
-            # but we use that value here in case we ever expand our history functionality.
-            vehicle_config_type = mission.vehicle_spec.veh_config_type
-            chassis_dims = BoundingBox.copy_with_defaults(
-                mission.vehicle_spec.dimensions,
-                VEHICLE_CONFIGS[vehicle_config_type].dimensions,
-            )
-        else:
-            # non-history agents can currently only control passenger vehicles.
-            vehicle_config_type = "passenger"
-            chassis_dims = VEHICLE_CONFIGS[vehicle_config_type].dimensions
+        chassis_dims = cls.agent_vehicle_dims(mission)
 
         if isinstance(mission.task, UTurn):
             if mission.task.initial_speed:
                 initial_speed = mission.task.initial_speed
 
         start = mission.start
-        start_pose = Pose.from_front_bumper(
-            front_bumper_position=np.array(start.position),
-            heading=start.heading,
-            length=chassis_dims.length,
-        )
+        if start.from_front_bumper:
+            start_pose = Pose.from_front_bumper(
+                front_bumper_position=np.array(start.position),
+                heading=start.heading,
+                length=chassis_dims.length,
+            )
+        else:
+            start_pose = Pose.from_center(start.position, start.heading)
 
         vehicle_color = (
             SceneColors.Agent.value if trainable else SceneColors.SocialAgent.value
@@ -378,11 +385,12 @@ class Vehicle:
 
         chassis = None
         # change this to dynamic_action_spaces later when pr merged
-        if (
-            agent_interface
-            and agent_interface.action in sim.dynamic_action_spaces
-            and not mission.vehicle_spec
-        ):
+        if agent_interface and agent_interface.action in sim.dynamic_action_spaces:
+            if mission.vehicle_spec:
+                logger = logging.getLogger(cls.__name__)
+                logger.warning(
+                    "setting vehicle dimensions on a AckermannChassis not yet supported"
+                )
             chassis = AckermannChassis(
                 pose=start_pose,
                 bullet_client=sim.bc,
@@ -533,6 +541,7 @@ class Vehicle:
         self._chassis.control(*args, **kwargs)
 
     def update_state(self, state: VehicleState, dt: float):
+        state.updated = True
         if not state.privileged:
             assert isinstance(self._chassis, BoxChassis)
             self.control(pose=state.pose, speed=state.speed, dt=dt)